backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pydantic[email]
email-validator
python-dotenv
psycopg2-binary
pydantic-settings
slowapi
cachetools
sentry-sdk[fastapi]
httpx
pytest
pytest-asyncio
